requirements.txt
# Core dependencies
python-osc>=1.8.0          # OSC communication with VRChat
PyQt6>=6.5.0                # Cross-platform GUI framework
numpy>=1.24.0               # Mathematical operations for proximity calculations
pyyaml>=6.0                 # Configuration file handling

# VR and 3D math dependencies
openvr>=1.21.4              # OpenVR/SteamVR integration
pyrr>=0.10.3                # 3D math library for position calculations

# Async and networking
asyncio                     # Async operations for real-time updates
websockets>=11.0            # Alternative communication method
aiofiles>=23.0              # Async file operations

# System integration
psutil>=5.9.0               # System monitoring and process management
platformdirs>=3.0.0        # Cross-platform directory handling

# Development dependencies (optional)
pytest>=7.0.0               # Testing framework
black>=23.0.0               # Code formatter
mypy>=1.0.0                 # Type checking
PyInstaller>=5.0.0          # Standalone executable builder
